requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
SpeechRecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
numpy==1.24.3 